remove old traverse stuff
